requirements.txt
aiofiles==0.4.0
Click==7.0
dataclasses==0.6
fastapi==0.30.0
h11==0.8.1
httptools==0.0.13
jinja2==2.10.1
MarkupSafe==1.1.1
pydantic==0.28
robotframework==3.1.2
SQLAlchemy==1.3.3
starlette==0.12.0
uvicorn==0.7.1
uvloop==0.12.2
websockets==7.0
